server/requirements.txt
#
# This file is autogenerated by pip-compile with python 3.8
# To update, run:
#
#    pip-compile
#
cachetools==5.2.0
    # via google-auth
certifi==2022.6.15
    # via requests
charset-normalizer==2.1.1
    # via requests
click==8.1.3
    # via flask
commonmark==0.9.1
    # via rich
cycler==0.11.0
    # via matplotlib
flask==2.2.2
    # via -r requirements.in
fonttools==4.37.0
    # via matplotlib
google-api-core==2.8.2
    # via google-api-python-client
google-api-python-client==2.57.0
    # via -r requirements.in
google-auth==2.11.0
    # via
    #   google-api-core
    #   google-api-python-client
    #   google-auth-httplib2
    #   google-auth-oauthlib
google-auth-httplib2==0.1.0
    # via google-api-python-client
google-auth-oauthlib==0.5.2
    # via -r requirements.in
googleapis-common-protos==1.56.4
    # via google-api-core
httplib2==0.20.4
    # via
    #   google-api-python-client
    #   google-auth-httplib2
idna==3.3
    # via requests
importlib-metadata==4.12.0
    # via flask
itsdangerous==2.1.2
    # via flask
jinja2==3.1.2
    # via flask
kiwisolver==1.4.4
    # via matplotlib
markupsafe==2.1.1
    # via
    #   jinja2
    #   werkzeug
matplotlib==3.5.3
    # via -r requirements.in
numpy==1.23.2
    # via
    #   matplotlib
    #   pandas
oauthlib==3.2.0
    # via requests-oauthlib
packaging==21.3
    # via matplotlib
pandas==1.4.3
    # via -r requirements.in
pillow==9.2.0
    # via matplotlib
protobuf==4.21.5
    # via
    #   google-api-core
    #   googleapis-common-protos
pyasn1==0.4.8
    # via
    #   pyasn1-modules
    #   rsa
pyasn1-modules==0.2.8
    # via google-auth
pygments==2.13.0
    # via rich
pyparsing==3.0.9
    # via
    #   httplib2
    #   matplotlib
    #   packaging
python-dateutil==2.8.2
    # via
    #   matplotlib
    #   pandas
python-dotenv==0.20.0
    # via -r requirements.in
pytz==2022.2.1
    # via pandas
requests==2.28.1
    # via
    #   google-api-core
    #   requests-oauthlib
requests-oauthlib==1.3.1
    # via google-auth-oauthlib
rich==12.5.1
    # via -r requirements.in
rsa==4.9
    # via google-auth
six==1.16.0
    # via
    #   google-auth
    #   google-auth-httplib2
    #   python-dateutil
typing-extensions==4.3.0
    # via rich
uritemplate==4.1.1
    # via google-api-python-client
urllib3==1.26.12
    # via requests
werkzeug==2.2.2
    # via flask
zipp==3.8.1
    # via importlib-metadata
